fahreza246d-update-delete-supplier-pada-controller-dan-model

+++ app/Models/Supplier.php
@@ -37,20 +37,24 @@ class Supplier extends Model
 
     public static function deleteSupplier($id)
     {
-    
         $supplier = static::where('supplier_id', $id)->first();
 
-    
         if (!$supplier) {
             return false;
         }
 
     
+        $purchaseOrderCount = DB::table('purchase_order')->where('supplier_id', $id)->count();
+    
+        if ($purchaseOrderCount > 0) {
+            return 'Supplier ini tidak bisa dihapus karena sudah memiliki purchase order';
+        }
+
         DB::table('purchase_order')->where('supplier_id', $id)->update(['supplier_id' => null]);
 
-    
         return $supplier->delete();
     }
+
 
 }
